Banking_System
This program is implementing a simple banking system with features to create an account, deposit money, and withdraw money. It is meant to run on an embedded system.

The main.c file contains the main function which acts as the driver code. It prompts the user to select an option from the menu and calls the appropriate function.

The Create_Account() function prompts the user to enter details like name, phone number, ID etc to create a new account. It validates the password and confirmation.

The Deposit() function asks for account login details like name, username, password. It then takes the deposit amount and prints confirmation.

Similarly, Withdraw() handles account withdrawal by taking login and withdrawal amount as input.

All the required header files are included. Functions are defined with proper prototypes. Input is taken using scanf and strings using fgets. Conditions are checked using if-else.

This embedded banking application is built to run on the target embedded platform. The C code uses software tools like compiler, linker etc during development and debugging. Overall it presents a basic but complete embedded application to manage bank accounts on an embedded system.